+++ projects/nb-trans/src/lib/components/nb-trans-subcontent/test/nb-trans-subcontent.component.spec.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectorRef, Component, TemplateRef, ViewChild } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NbTransTestingModule } from '../../testing';
-import { NbTransSubcontentComponent } from './nb-trans-subcontent.component';
+import { NbTransTestingModule } from '../../../testing';
+import { NbTransSubcontentComponent } from '../nb-trans-subcontent.component';
 
 @Component({
   selector: 'mock-tpl-ref',