+++ projects/nb-trans/src/lib/pipes/nb-sentence-item-type.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { NbValueTypeService } from '@bigbear713/nb-common';
-import { NbTransSentenceItemEnum } from '../constants';
+import { NbTransSentenceItem } from '../constants';
 import { INbTransSentencePart } from '../models';
 
 @Pipe({ name: 'nbSentenceItemType' })
@@ -12,11 +12,11 @@ export class NbSentenceItemTypePipe implements PipeTransform {
     let type: number | undefined;
 
     if (this.valueType.isString(value)) {
-      type = NbTransSentenceItemEnum.STR;
+      type = NbTransSentenceItem.STR;
     } else if (this.valueType.isNumber((value?.index))) {
       type = (Array.isArray(value.list) && value.list.length)
-        ? NbTransSentenceItemEnum.MULTI_COMP
-        : NbTransSentenceItemEnum.COMP;
+        ? NbTransSentenceItem.MULTI_COMP
+        : NbTransSentenceItem.COMP;
     }
 
     return type;