+++ projects/nb-trans/src/lib/pipes/test/nb-sentence-item-type.pipe.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { NbValueTypeService } from '@bigbear713/nb-common';
 import { isString } from 'lodash-es';
-import { NbTransSentenceItemEnum } from '../../constants';
+import { NbTransSentenceItem } from '../../constants';
 import { INbTransSentencePart } from '../../models';
 import { NbSentenceItemTypePipe } from '../nb-sentence-item-type.pipe';
 
@@ -20,9 +20,9 @@ describe('Pipe: NbSentenceItemType', () => {
   describe('#transform()', () => {
     [
       { params: undefined as any, expect: undefined },
-      { params: 'strContent', expect: NbTransSentenceItemEnum.STR },
-      { params: { index: 0, content: 'strContent', list: [] }, expect: NbTransSentenceItemEnum.COMP },
-      { params: { index: 0, content: '<0>str</0>', list: [{ index: 0, content: 'str', list: [] }] }, expect: NbTransSentenceItemEnum.MULTI_COMP },
+      { params: 'strContent', expect: NbTransSentenceItem.STR },
+      { params: { index: 0, content: 'strContent', list: [] }, expect: NbTransSentenceItem.COMP },
+      { params: { index: 0, content: '<0>str</0>', list: [{ index: 0, content: 'str', list: [] }] }, expect: NbTransSentenceItem.MULTI_COMP },
       { params: { index: undefined, content: 'strContent', list: [] } as unknown as INbTransSentencePart, expect: undefined },
       { params: { index: undefined, content: 'strContent', list: undefined } as unknown as INbTransSentencePart, expect: undefined },
     ].forEach(item => {