+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - feat: `angular`升级到`v15`; 
 - feat: `@bigbear713/nb-common`升级到`v15.0.0`; 
 - feat: `uuid`升级到`v9.0.0`;
-- feat: `tslib`升级到`v2.4.1`;
+- refactor: 优化nb-trans service和nb-trans pipe;
 
 <br/>